rm pipeline_description.json
This file was at the moment redundant with pyproject.toml, while also
not being duplicated in afids-auto-apply train. If we are to retain this
file, we should:

(1) Have it mirror the information from the pyproject.toml
(2) Either have it in a central location for both workflows, or have it
duplicated across the train and apply workflows.